Updated parameters container to Ordered_dict to preserve parameter order.
Moved value from an attribute to a property. In connection to this, I made the consequence of entering an illegal value an error instead of it just being ignored. This was chosen as the setter function is now hidden from the user, so they wouldnt necessarily know a check is being made behind the scences.

Updated tests to account for removal of set_value function.